Move secret routes to separate file
